Fix errors when adding a new sound with modified shape
If a new sound is created, it is not actually drawn right away; only a preview is drawn. This causes problems if refreshing the sound, because the refresh (or update?) cause the non-drawn sound to be added to canvas.sounds.placeables. Eventually, canvas.sounds layer is refreshed, and if there is a non-drawn sound object in placeables, it throws an error because the sound has no field to set its position. This is a work-around that continues to use preview sounds but cleans up after.