Orientation fixes for RegularPolygon
Check side can use orientation from origin to each vertex to determine the side. 

Use numSides to ensure all pointsBetween are captured for different shapes when a and b share a side.

When getting segment intersections, start with the side for a to ensure the intersections are in order of a to b.