🐛 fix|CachedEdges|Switch to using canvas.edges
Canvas.edges stores all possible edges for the scene. Best place to store cached wall edges. Use a custom type based on the source id so that the walls for that source can be used.

Works, but need to update on refresh, not update hook, so wall is dragged with the light.